backend/requirements.txt
Flask==3.0.3
mysql-connector-python==8.4.0
flask-cors==4.0.1